move two api to another resource
